Fix part of the issue with main thread tasks in tests
