+++ src/Common/Core/Test/Microsoft.Common.Core.Test.csproj
@@ -65,7 +65,6 @@
     <Compile Include="Fakes\Shell\TestFileDialog.cs" />
     <Compile Include="Fakes\Shell\TestIdleTimeService.cs" />
     <Compile Include="Fakes\Shell\TestImageService.cs" />
-    <Compile Include="Fakes\Shell\TestProgressDialog.cs" />
     <Compile Include="Fakes\Shell\TestUIServices.cs" />
     <Compile Include="Fixtures\ServiceManagerFixture.cs" />
     <Compile Include="Fixtures\ServiceManagerWithMefFixture.cs" />
