+++ src/UnitTests/Core/Impl/Microsoft.UnitTests.Core.csproj
@@ -27,6 +27,7 @@
     <Compile Include="..\..\..\GlobalAssemblyInfo.cs">
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="Threading\TestMainThread.cs" />
     <Compile Include="XUnit\TaskObserver.cs" />
     <Compile Include="XUnit\VsAssemblyLoaderAttribute.cs" />
     <Compile Include="XUnit\XunitTestEnvironment.cs" />
